requirements.txt
annotated-types==0.7.0
anyio==4.6.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
exceptiongroup==1.2.2
Faker==37.0.0
fastapi==0.115.2
h11==0.16.0
httptools==0.6.4
humanfriendly==10.0
ibm-cloud-sdk-core==3.22.0
ibm-code-engine-sdk==4.1.1
ibm-platform-services==0.59.0
idna==3.10
polyfactory==2.19.0
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
setuptools==78.1.1
six==1.16.0
sniffio==1.3.1
starlette==0.40.0
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.1
